+++ project_repo/hooks/pre_gen_project.py
@@ -6,17 +6,17 @@
 
     {# DevOps-specific prompt #}
     {% if is_devops %}
-        {{ cookiecutter.update({"organisation_name": prompt_user("DevOps Organisation name", "organisation_name")}) }}
-        {{ cookiecutter.update({"private_agent_name": prompt_user("DevOps Private agent name", "private_agent_name")}) }}
-        {{ cookiecutter.update({"package_feed": prompt_user("Package feed name", "package_feed")}) }}
+        {{ cookiecutter.update({"organisation_name": prompt_user("organisation_name")}) }}
+        {{ cookiecutter.update({"private_agent_name": prompt_user("private_agent_name")}) }}
+        {{ cookiecutter.update({"package_feed": prompt_user("package_feed")}) }}
         {{ cookiecutter.update({"azure_resource_group": "azure_resource_group"}) }}
         {{ cookiecutter.update({"azureml_workspace": "azureml_workspace"}) }}
     {% endif %}
 
     {# GitHub-specific prompts #}
     {% if is_github %}
-        {{ cookiecutter.update({"azure_resource_group": prompt_user("Azure resource group", "azure_resource_group")}) }}
-        {{ cookiecutter.update({"azureml_workspace": prompt_user("AzureML workspace", "azureml_workspace")}) }}
+        {{ cookiecutter.update({"azure_resource_group": prompt_user("azure_resource_group")}) }}
+        {{ cookiecutter.update({"azureml_workspace": prompt_user("azureml_workspace")}) }}
         {{ cookiecutter.update({"package_feed": "package_feed"}) }}
         {{ cookiecutter.update({"organisation_name": "organisation_name"}) }}
         {{ cookiecutter.update({"private_agent_name": "private_agent_name"}) }}